Made compile_filename file output optional
There are uses cases to get the QSS compiled string from a file without having to save it to a file.
This makes the `output_file` in `compile_filename` optional.
If undefined it will not save the output to a file and will just return it.
The documentation for `compile_filename` in the readme was duplicated so I removed the duplication.